Push common viz code between historybydate and historybytaxonomy up into the super class
